server/requirements.txt
Flask
flask-jwt-extended
flask-cors
mongoengine
python-dotenv
slack_sdk
boto3
requests